requirements.txt
alembic==1.14.1
biopython==1.83
catboost==1.2.7
category-encoders==2.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
fonttools==4.53.1
graphviz==0.20.3
greenlet==3.1.1
gseapy==1.1.4
idna==3.10
imbalanced-learn==0.12.3
importlib_metadata==8.5.0
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.41.1
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-venn==1.1.1
narwhals==1.27.1
numba==0.58.1
numpy==1.24.4
optuna==4.2.1
orfipy==0.0.4
packaging==24.1
pandas==2.0.3
patsy==0.5.6
pillow==10.4.0
plotly==6.0.0
psutil==6.0.0
pyahocorasick==2.1.0
pyfastx==2.1.0
pyparsing==3.1.4
pypinyin==0.53.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shap==0.44.1
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.38
statsmodels==0.14.1
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
xgboost==2.0.3
zipp==3.20.2
